// src/app/models/tutor.model.ts
export interface ITutor {
    id:                  number;
    nombreTutor:         string;
    apellidoTutor:       string;
    dniTutor:            string;
    celularDelTutor:     string;
    emailDelTutor:       string;
    direccionDelTutor:   string;
    fotoDniFrontalTutor: string;
    fotoDniDorsalTutor:  string;
    idAtleta:            string;
}